Allow reading CSRF token value from cookie named after csrf-param value
Currently we store the token in a cookie and on turbolinks:render, we
refresh the <meta name="csrf-token"> tag with the value from the cookie,
only to have Turbolinks read the value from that meta tag before
submitting forms. We can simply skip the <meta> tag dance and just have
Turbolinks read the value directly from the cookie.