turbo-frame: Progressively enhance target attribute
Since `<turbo-frame>` elements are a spiritual success to `<iframe>`
elements, there is an opportunity to build atop a concept with existing
history and understanding.

Replace the `data-turbo-frame` attribute with the `target` attribute,
progressively enhancing both [`<form target="...">`][form-target] and
[`<a target="...">`][a-target].

Instead of introducing a new `data-turbo-frame` attribute, layer on
additional behavior to a well understood attribute that determines which
`<iframe>` element on the page to submit the `<form>` element or
navigate the `<a>` element within.

In both cases, the `_top` value is a special case value:

> the topmost browsing context (the "highest" context that’s an ancestor
> of the current one). If no ancestors, behaves as _self.

This commit also modifies the check for `[data-turbo-frame="top"]` to
instead be a check for `[target="_top"]`.

To set the `target` attribute on [an `<a>` constructed with
`link_to`][link_to], pass it directly as `target:`.

To set the `target` attribute on [a `<form>` constructed with
`form_with`][form_with], pass it as part of the `html: { target: ... }`
option.

To set the `target` attribute on [a `<form>` constructed with
`button_to`][button_to], pass it as part of the `form: { target: ... }`
option.

[form-target]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/form#attr-target
[a-target]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a#attr-target
[link_to]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/UrlHelper.html#method-i-link_to
[form_with]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/FormHelper.html#method-i-form_with
[button_to]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/UrlHelper.html#method-i-button_to