Support Lazy-loading `<turbo-frame>` contents
Add support for declaring `[loading="lazy"]` (in the style of
[`iframe[loading="lazy"]`][iframe-loading] and
[`img[loading="lazy"]`][img-loading]). When both `src` and
`loading="lazy"` are declared on the frame, navigation will be deferred
until the element is visible in the document.

As an example, consider a `<details>`/`<summary>` pairing:

```html
<details>
  <summary>Lazy load</summary>

  <turbo-frame id="frame" src="/lazy-loaded.html" loading="lazy"></turbo-frame>
</details>
```

On initial load, the frame's contents will be empty. When the
`<summary>` element is clicked and the contents are visible, the
`<turbo-frame>` element will navigate to the URL specified in the `src`
element.

By default, `<turbo-frame>` elements that omit the `loading` attribute
will be treated as if they've declared `loading="eager"`.

The attribute (and corresponding `TurboFrameElement.loading` property
accessors) provide a means of controlling the loading strategy
declaratively without additional application-side JavaScript, while
still providing a means for application-side JavaScript to intervene if
necessary.

If visibility-based intersection is insufficient for an application's
need, there are various ways to control the loading strategy.

For example, consider an optimization to pre-load the frame when the
`<details>` element's disclosure `<summary>` button is hovered or
focused. First, consider HTML annotated with Stimulus attributes:

```html
<details data-controller="lazy-loader">
  <summary data-action="hover->lazy-loader#prefetch focus->lazy-loader#prefetch">Lazy load</summary>

  <turbo-frame id="frame" src="/lazy-loaded.html" loading="lazy" data-lazy-loader-target="frame"></turbo-frame>
</details>
```

Next, consider the corresponding `lazy-loader` Stimulus controller:

```js
import { Controller } from "stimulus"

export default class extends Controller {
  static targets = [ "frame" ]

  prefetch() {
    this.frameTarget.loading = "eager"
  }
}
```

[img-loading]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Img#attr-loading
[iframe-loading]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe#attr-loading