Avoid redefining things when using the asset pipeline in Safari
While using Safari with Turbo and Stimulus and a Stimulus controller
that also utilizes Turbo I am seeing these errors.

When Stimulus tries to import the controller, it seems as if it is
trying to import it twice causing these "already defined" errors.

While using Turbo and Stimulus through the asset pipeline
- Only define customeElements when they are not already
- Only define 'submitter' property if it isnt already defined on Event