Expand rendering capabilities of `<turbo-frame>`
Problem
---

Navigating a `<turbo-frame>` element always renders the new content by
updating the contents with the children from the response. This limits
the utility of `<turbo-frame>` in scenarios involving pagination driven
by "next" and "previous" page links.

Solution
---

Add support for `<turbo-frame rendering="...">`, `<a
data-turbo-rendering="...">`, `<form data-turbo-rendering="...">`, and
`<button data-turbo-rendering="...">` where the value of `[rendering]`
and `[data-turbo-rendering]` is one of the values that `<turbo-stream
action="...">` supports.

By default, `<turbo-frame>` elements will continue to render render with
the behavior equivalent to a `<turbo-stream action="update">` element.

This commit extends that support to also include the other actions:

* `after` will insert the contents of the response frame after the
  request frame
* `append` will extract the contents out of the response frame and
  append them into the request frame
* `before` will insert the contents of the response frame before the
  request frame
* `prepend` will extract the contents out of the response frame and
  append them into the request frame
* `replace` will extract the contents out of the response frame, remove
  the request frame, and inject the extracted contents in its place
  (conceptually similar to setting `outerHTML`)
* `remove` will remove the request frame, and ignore the contents of the
  response frame

This enables behaviors that might have been achievable with
`GET`-request powered Turbo Stream responses.

For example, in-place pagination could be achieved with
`rendering="prepend"` or `rendering="append"`:

```html
<!-- current HTML -->
<turbo-frame id="posts">
  <article id="article_1"><!-- contents --></article>
  <!-- articles 2-9 -->
  <article id="article_10"><!-- contents --></article>

  <a href="/posts?page=2" data-turbo-rendering="append">Next page</a>
</turbo-frame>

<!-- response HTML -->
<turbo-frame id="posts">
  <article id="article_11"><!-- contents --></article>

  <a href="/posts?page=3" data-turbo-rendering="append">Next page</a>
</turbo-frame>

<!-- HTML after the request -->
<turbo-frame id="posts">
  <article id="article_1"><!-- contents --></article>
  <!-- articles 2-9 -->
  <article id="article_10"><!-- contents --></article>
  <a href="/posts?page=2" data-turbo-rendering="append">Next page</a>
  <article id="article_11"><!-- contents --></article>

  <a href="/posts?page=3" data-turbo-rendering="append">Next page</a>
</turbo-frame>
```

Through the power of a CSS rules utilizing `:last-of-type`, we can hide
the pagination links:

```css
  #posts a              { display: none; }
  #posts a:last-of-type { display: block; }
```