Prevent unnecessary requests when restoring same-page visits with caching disabled.
Requests shouldn't be issued when navigating back/forward on same-page visits; the scroll position should just be restored. However, when caching is disabled Turbo does not have a scroll position value to restore to.

This change prioritizes checking for same-page visits to prevent unnecessary requests, and when restoring a scroll position, tries to scroll to the anchor if it exists. It's effectively the same outcome as before, without the request.