Support verb overrides through _method
Closes https://github.com/hotwired/turbo/issues/154.
Closes https://github.com/hotwired/turbo/pull/155.

---

To match the Rails idiom of override a form's `method` to support
requests that are not `GET` and `POST`, this commit adds test coverage
for existing functional behavior, along with some edge cases:

* considering a submitter's `formmethod` value over a `<input
  type="hidden" name="_method" value="...">` value

* considering a submitter's `formmethod` value over a competing
  submitter `name="_method" value="..."` pairing

To do so, change `FormSubmission.method` to be a read-only property
generated at construction time, and remove the property.

To read the `_method` requires destructively modifying the `FormData`
instance, so make sure we only do that once.

Co-authored-by: at-fffx <fangxing204@gmail.com>