Don't duplicate Form Submitter [name] param
When a `<form>` element's submitter declares a `[name]` attribute
without a `[value]` attribute, it is appended to the submission's
`FormData` instance twice. Due to a quirky difference between Safari and
Firefox + Chrome, the duplication _only_ occurs in Safari, which is a
browser that we are not running our functional test suite against.

In spite of that, this commit adds test coverage to ensure we don't
introduce a similar regression into those browsers.

Testing
---

Oddly enough, there are several occurrences of doubled-up `await await`
keywords, which this commit removes.