Prevent Streams from {pre,app}ending duplicates
In a circumstance where a `<form>` submission triggers a `turbo-stream`
response or a redirect back to the page with the new record on it, a
corresponding Stream update broadcast might occur in the background that
results in an `append` or `prepend` action.

In the case of the broadcast losing the race condition, update the
Stream renderer's page update to prevent prepending or appending
elements with duplicate `[id]` attributes.

Testing
---

Modify the test server to accept `action` and `target` values for
resulting Streams as request body parameters.