Support `data-turbo-action="..."` on Form Submits
The `Navigator` class translates `<form method="get">` submissions into
`proposeVisit` calls, skipping the rest of the `FormSubmission`
architecture.

This commit adds support for declaring `[data-turbo-action="replace"]`
(or `"restore"` or `"advance"`) as a [Visit Action][]. Prior to this
commit, specifying a `Visit` action was limited to an `<a>`
click-initiated `Visit` only.

This can be useful for a `GET` form driven by auto-submission driving
the current page's URL, like a Typeahead autocomplete `<input
type="search">` field or a bounding-box scrolling Map. In those
scenarios, a series of `"advance"` page visits would yield an
essentially broken "Back" and "Forward" navigation experience.

[Visit Action]: https://turbo.hotwire.dev/handbook/drive#application-visits