Prevent reloading pages when tapping same-page anchors (#298)
* Add test coverage for Turbo Drive and Skip Links

Closes https://github.com/hotwired/turbo/issues/42

Navigate with a skip link within our Functional test suite, assert that
the correct element is scrolled to, the page's Location path and hash
are correct, and that the initial tab stop occurs after the skipped-to
content.

* Discern URLs with blank hash vs no hash.
getAnchor returns an empty string for blank hashes (#) and undefined when no hash is set at all

* Prevent re-rendering when tapping same-page anchor

* Ensure correct focus when navigating to anchor

* Test page is not reloaded

* Fix navigating back to same page anchor

* Test navigating back to anchored location

* Tidy

* Fix Visit#isSamePage
when navigating back from an anchored location to a non-anchored location on the same page

* Prevent unnecessary requests
when restoring same-page visits with caching disabled.

Requests shouldn't be issued when navigating back/forward on same-page visits; the scroll position should just be restored. However, when caching is disabled Turbo does not have a scroll position value to restore to.

This change prioritizes checking for same-page visits to prevent unnecessary requests, and when restoring a scroll position, tries to scroll to the anchor if it exists. It's effectively the same outcome as before, without the request.

* API for native adapters to check for same-page anchors

* Trigger hashchange after same-page visits

Co-authored-by: Sean Doyle <sean.p.doyle24@gmail.com>
Co-authored-by: Özgür Uysal <ozguruysal@gmail.com>