Implement the ability to disable Drive by default.
Add a session property `drive` which is `true` by default.
Move `elementIsNavigable` back into session, rename to
`elementDriveEnabled` to better describe what it does, and adjust it to
use the new `drive` session property to determine if Drive should be
enabled or disabled by default.
Add `DriveTests` to test the basic behavior of Drive in the default
state and after `Turbo.session.drive` is set to false.