Replace `reloadable` attribute with `currentURL` property
As an alternative to leaking internal `FrameController` state to the DOM
as the `[reloadable]` attribute, this commit replaces those assignments
and guard clause checks with references to `currentURL` instead.

When responding to `<a>` clicks and `<form>` submissions, reset the
`<turbo-frame>` element delegates `currentURL = null`.

Adding a "public" (that is, package-internal) accessor to the
`FrameElementDelegate` feels less committal than an end-user facing
`[reloadable]` attribute, and achieves the same outcome.