Target must be in the document.body to be the event emitter (#403)
* Target must be in the document.body to be the event emitter

It's possible for this not to be true if the target element has been removed from the dom between when a form submission was made and the response is processed. This can happen as a race condition when the same turbo stream removal is triggered by both websocket and server response.

* Drop needless voids

* Use Node#isConnected

Also works on shadow doms!