Renderer: Extract `renderElement()`
For each renderer (`PageRenderer, `ErrorRenderer`, `FrameRenderer`)
extract their rendering logic into a class-`static-` property named
`renderElement()`. That function accepts two arguments: the first is the
current version of the element being rendered, the second is the new
version of the element being rendered.

The `PageRenderer` and `ErrorRenderer` types are genericized to accept
`HTMLBodyElement` instances. The `FrameRenderer` generics accept a
`FrameElement` instance. This means that the `renderElement()` function
shares the same class-level generics.