Frames: Ignore already cancelled `submit` events
Closes [hotwired/turbo#243][]

In tandem with [rails/rails#41960][] (which shipped as part of
`7.0.0.alpha2`), this commit modifies the `FormInterceptor` to ignore
`submit` events that have been cancelled.

Although it doesn't test the integration directly, the cancellation of
the `submit` event is what powers the ActiveStorage.js direct upload
asynchrony. Following a direct upload, ActiveStorage dispatches a
synthetic `submit` event, which the `FormInterceptor` will then respond
to.

[hotwired/turbo#243]: https://github.com/hotwired/turbo/issues/243
[rails/rails#41960]: https://github.com/rails/rails/pull/41960