Adds append_unless_duplicate and prepend_unless_duplicate
When appending or prepending content we want to control what
the action should do if there are already matching elements
on the page.
This is especially true when responding and broadcasting streams
contain the same data.

Additional actions added to skip the action if elements are
already present. In this case it would be recommended to use
the `replace` action.