Custom Actions: Export `StreamActions` module
Closes https://github.com/hotwired/turbo/pull/478
Closes https://github.com/hotwired/turbo/pull/477

As an alternative to continuing to introduce support for more varied and
specialized `<turbo-stream action="...">` values, this commit exports
the `StreamActions` module from the root-module in order to provide
applications with direct access.

Applications can add new actions or override existing actions, so long
as they adhere to the `StreamAction` interface: a function whose `this`
context is an instance of `StreamElement`.