Bug: `getUserMedia` doesn't work reliably with Turbo on some iOS versions
I haven't figured out exactly how to write a test for this. But pushing this as a replication example.

To replicate:

- [Run the test server](https://github.com/hotwired/turbo/blob/main/CONTRIBUTING.md)
- Use a tool like `ngrok` to make the test server accessible over the internet
- On your iPhone, using Safari, go to the test server `/src/tests/fixtures/rendering.html`
- Tap on "Page with webcam"
- On the next page, you'll get a prompt to permit access to the camera
- Tap "approve"
- On some iOS versions, you won't see a camera feed. On some you will.

On iOS 14.4.2 you don't see a camera feed (or at least, I don't). On iOS 15.0.2 you do (or at least, I do).

I'm not convinced this is is a Turbo issue, but I am creating it here in the hopes anyone has any ideas on how to fix it. If you navigate to `src/tests/fixtures/page_with_webcam.html` then the video feed shows, on every browser I've tested. It only breaks (on some phones) when you get there via a link.